トークのビューコントローラ廃止
