--- merry/merry/Modules/StoryList/Call/MerryTalkViewController.swift
@@ -1,53 +1,0 @@
-//
-//  MerryTalk.swift
-//  merry
-//
-//  Created by 大川恭平 on 2018/02/28.
-//  Copyright © 2018年 kazu. All rights reserved.
-//
-
-import UIKit
-import AudioToolbox
-import AVFoundation
-
-class MerryTalkViewController: UIViewController {
-
-    var audioPlayer = AVAudioPlayer()
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        //テストです
-        playVoice(fileName: "why", type: "mp3")
-        // Do any additional setup after loading the view.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-    
-    public func playVoice(fileName:String,type:String){
-        do{
-            let filePath = Bundle.main.path(forResource: fileName, ofType: type)
-            let audioPath = URL(fileURLWithPath: filePath!)
-            audioPlayer = try AVAudioPlayer(contentsOf: audioPath)
-            audioPlayer.numberOfLoops = 0
-            audioPlayer.prepareToPlay()
-            audioPlayer.play()
-        }catch{
-            print("error play merry voice")
-        }
-    }
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destinationViewController.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
